requirements.txt
flask==3.0.0
twilio==8.11.0
openai==1.10.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
